#10442 added Shorty to the web.xml (change didn't make it) and added resizing and jpeg
