+++ dotCMS/WEB-INF/web.xml
@@ -174,10 +174,6 @@
     <listener>
     	<listener-class>com.dotcms.listeners.SessionMonitor</listener-class>
     </listener>
-     <listener>
-        <display-name>ImageIO service provider loader/unloader</display-name>
-        <listener-class>com.twelvemonkeys.servlet.image.IIOProviderContextListener</listener-class>
-    </listener>
 	<!-- this must be the last listener -->
 	<listener>
 	  <listener-class>com.dotmarketing.listeners.HibernateSessionsListener</listener-class>
@@ -220,8 +216,8 @@
 		<servlet-class>com.dotmarketing.servlets.SpeedyAssetServlet</servlet-class>
 	</servlet>
 	<servlet>
-		<servlet-name>SpeedyAssetServlet2</servlet-name>
-		<servlet-class>com.dotmarketing.servlets.SpeedyAssetServlet2</servlet-class>
+		<servlet-name>ShortyServlet</servlet-name>
+		<servlet-class>com.dotmarketing.servlets.ShortyServlet</servlet-class>
 	</servlet>
 	
 	<servlet>
@@ -497,7 +493,7 @@
 		<url-pattern>/dotAsset/*</url-pattern>
 	</servlet-mapping>
 	<servlet-mapping>
-		<servlet-name>SpeedyAssetServlet2</servlet-name>
+		<servlet-name>ShortyServlet</servlet-name>
 		<url-pattern>/dA/*</url-pattern>
 	</servlet-mapping>
 	<servlet-mapping>
